refactor(validation): merge 22 modules into 4 focused subpackages
Reorganize validation from flat 22-module structure into focused subpackages:
- validation/docs/: docstrings, doc_config, doc_policy
- validation/code/: type_aliases, complexity, mypy_per_file
- validation/tiers/: tier_labels, cli_tier_docs
- validation/skills/: skill_catalog, repo_analyze_skills, skill_merge_method

Each subpackage exposes single main() entry point. Reduces module count from 22 to ~12, clarifies domain boundaries, consolidates shared CLI boilerplate.

Update pyproject.toml console script entry points and test imports.

Closes #1446

Implemented-By: claude-opus-4-8
