+++ hephaestus/validation/__init__.py
@@ -4,17 +4,19 @@ from hephaestus.validation.audit import (
     filter_audit_results,
     severity_label,
 )
-from hephaestus.validation.complexity import check_max_complexity
+from hephaestus.validation.code.complexity import check_max_complexity
+from hephaestus.validation.code.mypy_per_file import check_mypy_per_file
+from hephaestus.validation.code.type_aliases import detect_shadowing, is_shadowing_pattern
 from hephaestus.validation.config_lint import ConfigLinter
 from hephaestus.validation.coverage import check_coverage, parse_coverage_report
-from hephaestus.validation.doc_config import check_doc_config_consistency
-from hephaestus.validation.doc_policy import (
+from hephaestus.validation.docs.doc_config import check_doc_config_consistency
+from hephaestus.validation.docs.doc_policy import (
     Finding as DocPolicyFinding,
     Severity,
     scan_file as scan_doc_policy,
     scan_repository as scan_doc_policy_repository,
 )
-from hephaestus.validation.docstrings import (
+from hephaestus.validation.docs.docstrings import (
     FragmentFinding,
     is_genuine_fragment,
     scan_file as scan_docstrings,
@@ -36,7 +38,6 @@ from hephaestus.validation.markdown import (
     validate_readme,
     validate_relative_link,
 )
-from hephaestus.validation.mypy_per_file import check_mypy_per_file
 from hephaestus.validation.python_version import (
     check_ci_matrix_coverage,
     check_pixi_python_ceiling,
@@ -63,8 +64,10 @@ from hephaestus.validation.test_structure import (
     check_test_directory_mirrors,
     check_test_structure,
 )
-from hephaestus.validation.tier_labels import TierLabelFinding, scan_repository as scan_tier_labels
-from hephaestus.validation.type_aliases import detect_shadowing, is_shadowing_pattern
+from hephaestus.validation.tiers.tier_labels import (
+    TierLabelFinding,
+    scan_repository as scan_tier_labels,
+)
 
 __all__ = [
     "CodeBlock",
