+++ hephaestus/validation/doc_config.py
@@ -1,539 +1,39 @@
-"""Enforce consistency between documentation metric values and authoritative config sources.
+"""Backward-compatibility shim.
 
-Checks that values documented in CLAUDE.md and README.md match what is configured in
-``pyproject.toml``:
-
-1. Coverage threshold in CLAUDE.md AND docs/DEFINITION_OF_DONE.md matches ``fail_under``
-   in ``[tool.coverage.report]``.
-2. ``--cov=<path>`` in README.md matches ``addopts`` in ``[tool.pytest.ini_options]``.
-3. If ``--cov-fail-under=N`` is present in ``addopts``, it must match ``fail_under`` in
-   ``[tool.coverage.report]``.  Absent is OK — ``[tool.coverage.report].fail_under`` is
-   the single source of truth.
-4. Test count in README.md is within 10% of actual ``pytest --collect-only`` count.
-
-Usage::
-
-    hephaestus-check-doc-config
-    hephaestus-check-doc-config --repo-root /path/to/repo --verbose
-
-Exit codes:
-    0  All checks pass
-    1  One or more checks failed
+Canonical implementation: :mod:`hephaestus.validation.docs.doc_config`.
+Retained so existing import sites and the ``hephaestus-check-doc-config``
+console script need no changes.
 """
 
-from __future__ import annotations
-
-import re
-import subprocess
 import sys
-from pathlib import Path
-from typing import Any, cast
 
-from hephaestus.cli.utils import create_validation_parser, format_output, resolve_repo_root
-from hephaestus.io.toml import import_tomllib
-from hephaestus.utils.helpers import NETWORK_TIMEOUT
-
-_tomllib = import_tomllib()
-
-
-def _load_pyproject(repo_root: Path) -> dict[str, Any]:
-    """Load ``pyproject.toml`` using tomllib/tomli.
-
-    Args:
-        repo_root: Root directory of the repository.
-
-    Returns:
-        Parsed TOML data as a nested dict.
-
-    Raises:
-        SystemExit: With code 1 if the file is missing, unreadable, or tomllib is
-            unavailable.
-
-    """
-    pyproject_path = repo_root / "pyproject.toml"
-    if not pyproject_path.is_file():
-        print(f"ERROR: pyproject.toml not found: {pyproject_path}", file=sys.stderr)
-        sys.exit(1)
-
-    if _tomllib is None:
-        print(
-            "ERROR: tomllib/tomli is required to parse pyproject.toml. "
-            "Install tomli for Python < 3.11: pip install tomli",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-
-    try:
-        with open(pyproject_path, "rb") as f:
-            return cast(dict[str, Any], _tomllib.load(f))
-    except Exception as exc:
-        print(f"ERROR: Could not parse {pyproject_path}: {exc}", file=sys.stderr)
-        sys.exit(1)
-
-
-def load_coverage_threshold(repo_root: Path) -> int:
-    """Read ``fail_under`` from ``[tool.coverage.report]`` in ``pyproject.toml``.
-
-    Args:
-        repo_root: Repository root containing ``pyproject.toml``.
-
-    Returns:
-        Integer value of ``fail_under``.
-
-    Raises:
-        SystemExit: If the key is absent or ``pyproject.toml`` is unreadable.
-
-    """
-    data = _load_pyproject(repo_root)
-    try:
-        threshold = data["tool"]["coverage"]["report"]["fail_under"]
-    except KeyError:
-        print(
-            "ERROR: [tool.coverage.report].fail_under not found in pyproject.toml",
-            file=sys.stderr,
-        )
-        sys.exit(1)
-    return int(threshold)
-
-
-def extract_cov_path(repo_root: Path) -> str:
-    """Read the ``--cov=<path>`` value from ``[tool.pytest.ini_options].addopts``.
-
-    Args:
-        repo_root: Repository root containing ``pyproject.toml``.
-
-    Returns:
-        Package path string (e.g. ``"hephaestus"``).
-
-    Raises:
-        SystemExit: If the key is missing or no ``--cov=`` flag is found.
-
-    """
-    data = _load_pyproject(repo_root)
-    addopts = data.get("tool", {}).get("pytest", {}).get("ini_options", {}).get("addopts", [])
-    addopts_items: list[str] = addopts.split() if isinstance(addopts, str) else list(addopts)
-
-    for item in addopts_items:
-        m = re.match(r"^--cov=(.+)$", item)
-        if m:
-            return m.group(1)
-
-    print(
-        "ERROR: No --cov=<path> found in [tool.pytest.ini_options].addopts",
-        file=sys.stderr,
-    )
-    sys.exit(1)
-
-
-def extract_cov_fail_under_from_addopts(repo_root: Path) -> int | None:
-    """Read ``--cov-fail-under=N`` from ``[tool.pytest.ini_options].addopts``, if present.
-
-    Args:
-        repo_root: Repository root containing ``pyproject.toml``.
-
-    Returns:
-        Integer threshold if the flag is present, ``None`` otherwise.
-
-    """
-    data = _load_pyproject(repo_root)
-    addopts = data.get("tool", {}).get("pytest", {}).get("ini_options", {}).get("addopts", [])
-    addopts_items: list[str] = addopts.split() if isinstance(addopts, str) else list(addopts)
-
-    for item in addopts_items:
-        m = re.match(r"^--cov-fail-under=(\d+)$", item)
-        if m:
-            return int(m.group(1))
-    return None
-
-
-def check_claude_md_threshold(repo_root: Path, expected: int) -> list[str]:
-    """Check that CLAUDE.md documents the correct coverage threshold.
-
-    Searches for ``<N>%+ test coverage`` (or ``<N>% test coverage``) and verifies
-    the integer matches *expected*.
-
-    Args:
-        repo_root: Repository root.
-        expected: Authoritative threshold value from ``pyproject.toml``.
-
-    Returns:
-        List of error strings (empty if all checks pass).
-
-    """
-    claude_md = repo_root / "CLAUDE.md"
-    if not claude_md.exists():
-        return [f"CLAUDE.md not found at {claude_md}"]
-
-    text = claude_md.read_text(encoding="utf-8")
-    matches = re.findall(r"(\d+)%\+?\s+test coverage", text)
-    if not matches:
-        return [
-            "CLAUDE.md: No coverage threshold mention found "
-            "(expected pattern: '<N>%+ test coverage')"
-        ]
-
-    errors: list[str] = []
-    for raw in matches:
-        found = int(raw)
-        if found != expected:
-            errors.append(
-                f"CLAUDE.md: Coverage threshold mismatch — "
-                f"CLAUDE.md says {found}%, pyproject.toml says {expected}%"
-            )
-    return errors
-
-
-def check_dod_threshold(repo_root: Path, expected: int) -> list[str]:
-    """Check that docs/DEFINITION_OF_DONE.md documents the correct coverage threshold.
-
-    The DoD references the gate in two distinct phrasings that differ from
-    CLAUDE.md's: ``--cov-fail-under=<N>`` (row 10) and ``drops total under <N>%``
-    (row F2). Both must match the authoritative *expected* value. The match is
-    anchored on these gate-specific shapes rather than a free ``<N>%`` scan so it
-    does not false-positive on unrelated numbers (the 10% tolerance, Python 3.10).
-
-    Args:
-        repo_root: Repository root.
-        expected: Authoritative threshold value from ``pyproject.toml``.
-
-    Returns:
-        List of error strings (empty if all checks pass).
-
-    """
-    dod = repo_root / "docs" / "DEFINITION_OF_DONE.md"
-    if not dod.exists():
-        return [f"DEFINITION_OF_DONE.md not found at {dod}"]
-
-    text = dod.read_text(encoding="utf-8")
-    matches = re.findall(r"--cov-fail-under=(\d+)", text)
-    matches += re.findall(r"drops total under (\d+)%", text)
-    if not matches:
-        return [
-            "DEFINITION_OF_DONE.md: No coverage threshold mention found "
-            "(expected '--cov-fail-under=<N>' or 'drops total under <N>%')"
-        ]
-
-    errors: list[str] = []
-    for raw in matches:
-        found = int(raw)
-        if found != expected:
-            errors.append(
-                f"DEFINITION_OF_DONE.md: Coverage threshold mismatch — "
-                f"DEFINITION_OF_DONE.md says {found}%, pyproject.toml says {expected}%"
-            )
-    return errors
-
-
-def check_readme_cov_path(repo_root: Path, expected_path: str) -> list[str]:
-    """Check that all ``--cov=<path>`` occurrences in README.md match *expected_path*.
-
-    Args:
-        repo_root: Repository root.
-        expected_path: Authoritative ``--cov`` path from ``pyproject.toml``.
-
-    Returns:
-        List of error strings (empty if all checks pass or README has no ``--cov``).
-
-    """
-    readme = repo_root / "README.md"
-    if not readme.exists():
-        return [f"README.md not found at {readme}"]
-
-    text = readme.read_text(encoding="utf-8")
-    occurrences = re.findall(r"--cov=(\S+)", text)
-    if not occurrences:
-        return []
-
-    errors: list[str] = []
-    for path in occurrences:
-        if path != expected_path:
-            errors.append(
-                f"README.md: --cov path mismatch — "
-                f"README.md has '--cov={path}', pyproject.toml uses '--cov={expected_path}'"
-            )
-    return errors
-
-
-def check_addopts_cov_fail_under(repo_root: Path, expected: int) -> list[str]:
-    """Check that ``--cov-fail-under`` in addopts matches ``fail_under`` (if present).
-
-    If absent, the check passes — ``[tool.coverage.report].fail_under`` is the single
-    source of truth and pytest-cov reads it directly.
-
-    Args:
-        repo_root: Repository root.
-        expected: Authoritative threshold from ``[tool.coverage.report].fail_under``.
-
-    Returns:
-        List of error strings (empty if consistent or flag is absent).
-
-    """
-    addopts_threshold = extract_cov_fail_under_from_addopts(repo_root)
-    if addopts_threshold is None:
-        return []
-    if addopts_threshold != expected:
-        return [
-            f"pyproject.toml: --cov-fail-under mismatch — "
-            f"addopts has --cov-fail-under={addopts_threshold}, "
-            f"but [tool.coverage.report].fail_under={expected}"
-        ]
-    return []
-
-
-def collect_actual_test_count(repo_root: Path) -> int | None:
-    """Run ``pytest --collect-only -q`` and return the number of collected tests.
-
-    Args:
-        repo_root: Repository root where pytest should be invoked.
-
-    Returns:
-        Integer test count, or ``None`` if collection fails or is unparseable.
-
-    """
-    try:
-        result = subprocess.run(
-            [sys.executable, "-m", "pytest", "--collect-only", "-q", "tests/"],
-            cwd=repo_root,
-            capture_output=True,
-            text=True,
-            timeout=NETWORK_TIMEOUT,
-        )
-    except (FileNotFoundError, OSError, subprocess.TimeoutExpired):
-        return None
-
-    output = result.stdout + result.stderr
-    m = re.search(r"(\d+)\s+(?:tests?\s+)?(?:selected|collected)", output)
-    if m:
-        count = int(m.group(1))
-        return count if count > 0 else None
-    return None
-
-
-def check_readme_test_count(
-    repo_root: Path, actual_count: int, tolerance: float = 0.10
-) -> list[str]:
-    """Check that test count claims in README.md are within *tolerance* of *actual_count*.
-
-    Searches for patterns like ``3,500+ tests`` or ``3172 tests``.
-
-    Args:
-        repo_root: Repository root.
-        actual_count: Authoritative test count from ``pytest --collect-only``.
-        tolerance: Fractional tolerance (default 0.10 = 10%).
-
-    Returns:
-        List of error strings (empty if consistent or no hardcoded count found).
-
-    """
-    readme = repo_root / "README.md"
-    if not readme.exists():
-        return [f"README.md not found at {readme}"]
-
-    text = readme.read_text(encoding="utf-8")
-    raw_matches = re.findall(r"(\d[\d,]*)\+?\s+tests?", text, re.IGNORECASE)
-    if not raw_matches:
-        return []
-
-    errors: list[str] = []
-    for raw in raw_matches:
-        doc_count = int(raw.replace(",", ""))
-        if abs(doc_count - actual_count) / actual_count > tolerance:
-            errors.append(
-                f"README.md: Test count mismatch — "
-                f"README.md says {doc_count}, actual pytest count is {actual_count} "
-                f"(tolerance: {int(tolerance * 100)}%)"
-            )
-    return errors
-
-
-def check_doc_config_consistency(
-    repo_root: Path,
-    verbose: bool = False,
-    skip_test_count: bool = False,
-) -> int:
-    """Run all doc/config consistency checks and return an exit code.
-
-    Checks:
-    1. CLAUDE.md AND docs/DEFINITION_OF_DONE.md coverage threshold vs
-       ``[tool.coverage.report].fail_under``.
-    2. README.md ``--cov=<path>`` vs ``[tool.pytest.ini_options].addopts``.
-    3. ``--cov-fail-under`` in addopts (if present) vs ``fail_under``.
-    4. README.md hardcoded test count vs ``pytest --collect-only`` (skipped if
-       *skip_test_count* is True or pytest is unavailable).
-
-    Args:
-        repo_root: Repository root containing ``pyproject.toml``.
-        verbose: If True, print passing check names as well.
-        skip_test_count: If True, skip the live ``pytest --collect-only`` check.
-
-    Returns:
-        0 if all checks pass, 1 if any fail.
-
-    """
-    all_errors: list[str] = []
-
-    expected_threshold = load_coverage_threshold(repo_root)
-    all_errors.extend(_run_threshold_check(repo_root, expected_threshold, verbose))
-    all_errors.extend(_run_cov_path_check(repo_root, verbose))
-    all_errors.extend(_run_addopts_check(repo_root, expected_threshold, verbose))
-
-    if not skip_test_count:
-        all_errors.extend(_run_test_count_check(repo_root, verbose))
-
-    if all_errors:
-        for error in all_errors:
-            print(error, file=sys.stderr)
-        print(
-            f"\nFound {len(all_errors)} doc/config consistency violation(s).",
-            file=sys.stderr,
-        )
-        return 1
-    return 0
-
-
-def _run_threshold_check(repo_root: Path, expected: int, verbose: bool) -> list[str]:
-    """Run CLAUDE.md and DEFINITION_OF_DONE.md coverage threshold checks.
-
-    Args:
-        repo_root: Repository root.
-        expected: Expected threshold integer.
-        verbose: Print pass message if True.
-
-    Returns:
-        List of error strings (empty if passing).
-
-    """
-    errors = check_claude_md_threshold(repo_root, expected)
-    errors += check_dod_threshold(repo_root, expected)
-    if not errors and verbose:
-        print(
-            f"PASS: CLAUDE.md and DEFINITION_OF_DONE.md coverage threshold "
-            f"match pyproject.toml ({expected}%)"
-        )
-    return errors
-
-
-def _run_cov_path_check(repo_root: Path, verbose: bool) -> list[str]:
-    """Run README.md --cov path check and print verbose result.
-
-    Args:
-        repo_root: Repository root.
-        verbose: Print pass message if True.
-
-    Returns:
-        List of error strings (empty if passing).
-
-    """
-    expected_cov_path = extract_cov_path(repo_root)
-    errors = check_readme_cov_path(repo_root, expected_cov_path)
-    if not errors and verbose:
-        print(f"PASS: README.md --cov path matches pyproject.toml (--cov={expected_cov_path})")
-    return errors
-
-
-def _run_addopts_check(repo_root: Path, expected: int, verbose: bool) -> list[str]:
-    """Run addopts --cov-fail-under consistency check and print verbose result.
-
-    Args:
-        repo_root: Repository root.
-        expected: Expected threshold integer.
-        verbose: Print pass message if True.
-
-    Returns:
-        List of error strings (empty if passing).
-
-    """
-    errors = check_addopts_cov_fail_under(repo_root, expected)
-    if not errors and verbose:
-        addopts_val = extract_cov_fail_under_from_addopts(repo_root)
-        if addopts_val is not None:
-            print(
-                f"PASS: addopts --cov-fail-under matches "
-                f"[tool.coverage.report].fail_under ({expected}%)"
-            )
-        else:
-            print(
-                f"PASS: No --cov-fail-under in addopts — "
-                f"[tool.coverage.report].fail_under ({expected}%) is single source of truth"
-            )
-    return errors
-
-
-def _run_test_count_check(repo_root: Path, verbose: bool) -> list[str]:
-    """Run pytest test-count check and print verbose result.
-
-    Args:
-        repo_root: Repository root.
-        verbose: Print pass/skip message if True.
-
-    Returns:
-        List of error strings (empty if passing or skipped).
-
-    """
-    actual_count = collect_actual_test_count(repo_root)
-    if actual_count is None:
-        if verbose:
-            print("SKIP: Could not collect actual test count (pytest unavailable)")
-        return []
-    errors = check_readme_test_count(repo_root, actual_count)
-    if not errors and verbose:
-        print(f"PASS: README.md test count is within 10% of actual ({actual_count})")
-    return errors
-
-
-def main() -> int:
-    """CLI entry point for doc/config consistency checking.
-
-    Returns:
-        Exit code (0 if all checks pass, 1 if any fail).
-
-    """
-    parser = create_validation_parser(
-        "Enforce consistency between doc metric values and pyproject.toml",
-        epilog="Example: %(prog)s --repo-root /path/to/repo --verbose",
-    )
-    parser.add_argument(
-        "--verbose",
-        "-v",
-        action="store_true",
-        help="Print passing check names",
-    )
-    parser.add_argument(
-        "--skip-test-count",
-        action="store_true",
-        help="Skip the live pytest --collect-only test count check",
-    )
-
-    args = parser.parse_args()
-    repo_root = resolve_repo_root(args)
-
-    if args.json:
-        expected_threshold = load_coverage_threshold(repo_root)
-        all_errors: list[str] = []
-        all_errors.extend(check_claude_md_threshold(repo_root, expected_threshold))
-        all_errors.extend(check_dod_threshold(repo_root, expected_threshold))
-        cov_path = extract_cov_path(repo_root)
-        all_errors.extend(check_readme_cov_path(repo_root, cov_path))
-        all_errors.extend(check_addopts_cov_fail_under(repo_root, expected_threshold))
-        if not args.skip_test_count:
-            actual_count = collect_actual_test_count(repo_root)
-            if actual_count is not None:
-                all_errors.extend(check_readme_test_count(repo_root, actual_count))
-        exit_code = 0 if not all_errors else 1
-        report = {
-            "expected_threshold": expected_threshold,
-            "errors": all_errors,
-            "exit_code": exit_code,
-            "passed": not all_errors,
-        }
-        print(format_output(report, "json"))
-        return exit_code
-
-    return check_doc_config_consistency(
-        repo_root=repo_root,
-        verbose=args.verbose,
-        skip_test_count=args.skip_test_count,
-    )
+from hephaestus.validation.docs.doc_config import (
+    check_addopts_cov_fail_under as check_addopts_cov_fail_under,
+    check_claude_md_threshold as check_claude_md_threshold,
+    check_doc_config_consistency as check_doc_config_consistency,
+    check_dod_threshold as check_dod_threshold,
+    check_readme_cov_path as check_readme_cov_path,
+    check_readme_test_count as check_readme_test_count,
+    collect_actual_test_count as collect_actual_test_count,
+    extract_cov_fail_under_from_addopts as extract_cov_fail_under_from_addopts,
+    extract_cov_path as extract_cov_path,
+    load_coverage_threshold as load_coverage_threshold,
+    main as main,
+)
+
+__all__ = [
+    "check_addopts_cov_fail_under",
+    "check_claude_md_threshold",
+    "check_doc_config_consistency",
+    "check_dod_threshold",
+    "check_readme_cov_path",
+    "check_readme_test_count",
+    "collect_actual_test_count",
+    "extract_cov_fail_under_from_addopts",
+    "extract_cov_path",
+    "load_coverage_threshold",
+    "main",
+]
 
 
 if __name__ == "__main__":
