+++ pyproject.toml
@@ -117,26 +117,26 @@ hephaestus-download-dataset = "hephaestus.datasets.downloader:main"
 hephaestus-check-python-version = "hephaestus.validation.python_version:main"
 hephaestus-check-test-structure = "hephaestus.validation.test_structure:main"
 hephaestus-check-coverage = "hephaestus.validation.coverage:main"
-hephaestus-check-complexity = "hephaestus.validation.complexity:main"
+hephaestus-check-complexity = "hephaestus.validation.code.complexity:main"
 hephaestus-filter-audit = "hephaestus.validation.audit:main"
 hephaestus-validate-schemas = "hephaestus.validation.schema:main"
 hephaestus-validate-links = "hephaestus.validation.markdown:main"
 hephaestus-check-readmes = "hephaestus.validation.markdown:check_readmes_main"
-hephaestus-check-skill-catalog = "hephaestus.validation.skill_catalog:main"
-hephaestus-check-repo-analyze-skills = "hephaestus.validation.repo_analyze_skills:main"
-hephaestus-check-type-aliases = "hephaestus.validation.type_aliases:main"
-hephaestus-check-docstrings = "hephaestus.validation.docstrings:main"
-hephaestus-check-tier-labels = "hephaestus.validation.tier_labels:main"
+hephaestus-check-skill-catalog = "hephaestus.validation.skills.skill_catalog:main"
+hephaestus-check-repo-analyze-skills = "hephaestus.validation.skills.repo_analyze_skills:main"
+hephaestus-check-type-aliases = "hephaestus.validation.code.type_aliases:main"
+hephaestus-check-docstrings = "hephaestus.validation.docs.docstrings:main"
+hephaestus-check-tier-labels = "hephaestus.validation.tiers.tier_labels:main"
 hephaestus-fix-markdown = "hephaestus.markdown.fixer:main"
-hephaestus-audit-doc-policy = "hephaestus.validation.doc_policy:main"
+hephaestus-audit-doc-policy = "hephaestus.validation.docs.doc_policy:main"
 hephaestus-check-version-consistency = "hephaestus.version.consistency:check_version_consistency_main"
 hephaestus-coredump-handler = "hephaestus.forensics.coredump_handler:main"
 hephaestus-run-under-gdb = "hephaestus.forensics.gdb_runner:main"
 hephaestus-check-package-versions = "hephaestus.version.consistency:check_package_versions_main"
 hephaestus-bump-version = "hephaestus.version.consistency:bump_version_main"
-hephaestus-check-doc-config = "hephaestus.validation.doc_config:main"
+hephaestus-check-doc-config = "hephaestus.validation.docs.doc_config:main"
 hephaestus-check-stale-scripts = "hephaestus.validation.stale_scripts:main"
-hephaestus-mypy-each-file = "hephaestus.validation.mypy_per_file:main"
+hephaestus-mypy-each-file = "hephaestus.validation.code.mypy_per_file:main"
 hephaestus-check-links = "hephaestus.markdown.link_fixer:main"
 hephaestus-validate-anchors = "hephaestus.markdown.anchors:main"
 hephaestus-scaffold-subpackage = "hephaestus.scripts_lib.scaffold_subpackage:main"
@@ -149,7 +149,7 @@ hephaestus-validate-workflow-checkout = "hephaestus.ci.workflows:validate_workfl
 hephaestus-github-stats = "hephaestus.github.stats:main"
 hephaestus-agent-stats = "hephaestus.agents.stats:main"
 hephaestus-validate-agents = "hephaestus.agents.frontmatter:validate_agents_main"
-hephaestus-check-cli-tier-docs = "hephaestus.validation.cli_tier_docs:main"
+hephaestus-check-cli-tier-docs = "hephaestus.validation.tiers.cli_tier_docs:main"
 hephaestus-check-api-table-docs = "hephaestus.validation.api_table_docs:main"
 
 [project.urls]
