+++ tests/unit/validation/test_doc_config.py
@@ -267,7 +267,7 @@ class TestCollectActualTestCount:
 
     def test_pytest_collection_passes_timeout(self, tmp_path: Path) -> None:
         """The pytest --collect-only subprocess is bounded (#684)."""
-        with patch("hephaestus.validation.doc_config.subprocess.run") as mock_run:
+        with patch("hephaestus.validation.docs.doc_config.subprocess.run") as mock_run:
             mock_run.return_value = MagicMock(stdout="5 tests collected\n", stderr="")
             collect_actual_test_count(tmp_path)
         assert mock_run.call_args.kwargs["timeout"] == NETWORK_TIMEOUT
@@ -275,7 +275,7 @@ class TestCollectActualTestCount:
     def test_returns_none_on_timeout(self, tmp_path: Path) -> None:
         """A hung pytest collection degrades to None instead of hanging (#684)."""
         with patch(
-            "hephaestus.validation.doc_config.subprocess.run",
+            "hephaestus.validation.docs.doc_config.subprocess.run",
             side_effect=subprocess.TimeoutExpired(cmd="pytest", timeout=120),
         ):
             assert collect_actual_test_count(tmp_path) is None
